Add files via upload
We used a pandas library to read in our dataset.  We then created a Function that analyzed the "Review Text" column, and we calculated a Sentiment Value for each review. Then, after removing missing values, we made a NEW column in the data frame to show/hold our Sentiment Value!